+++ service-worker.js
@@ -1,6 +1,6 @@
-const APP_VERSION = "4.6.2";
+const APP_VERSION = "4.6.3";
 const CACHE_PREFIX = "entenarchiv-shell-";
-const CACHE_NAME = `${CACHE_PREFIX}v4-6-2`;
+const CACHE_NAME = `${CACHE_PREFIX}v4-6-3`;
 
 const CORE_SHELL = Object.freeze([
   "./index.html",